requirements.txt
pytest>=7.0.0
allure-pytest>=2.13.0
pywinauto>=0.6.8
PyYAML>=6.0
